+++ packages/pptxgenjs-std/src/index.ts
@@ -1,2 +1,4 @@
 export * from './layout'
 export * from './charts'
+export * from './text'
+export * from './tables'